// src/app/model/MenuItem.model.ts
export class MenuItem {
    menuItemId: number =0 ;
    menuItemName: string = '';
    menuItemPath: string = '' ;
    menuItemDescription: string = '';
    menuItemStatus: string;
    createdDateTime: string;
    modifiedDateTime:string;
    createdBy: string;
    modifiedBy: string;
    createdByEmailId: string;
    modifiedByEmailId: string;
}